+++ test/deployframework/framework.go
@@ -9,7 +9,7 @@ import (
 
 	olmclientv1 "github.com/operator-framework/operator-lifecycle-manager/pkg/api/client/clientset/versioned/typed/operators/v1"
 	olmclientv1alpha1 "github.com/operator-framework/operator-lifecycle-manager/pkg/api/client/clientset/versioned/typed/operators/v1alpha1"
-	apiextclientv1beta1 "k8s.io/apiextensions-apiserver/pkg/client/clientset/clientset/typed/apiextensions/v1beta1"
+	apiextclientv1 "k8s.io/apiextensions-apiserver/pkg/client/clientset/clientset/typed/apiextensions/v1"
 	meta "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/client-go/kubernetes"
 	"k8s.io/client-go/rest"
@@ -58,7 +58,7 @@ type DeployFramework struct {
 	Logger            logrus.FieldLogger
 	Config            *rest.Config
 	Client            kubernetes.Interface
-	APIExtClient      apiextclientv1beta1.CustomResourceDefinitionsGetter
+	APIExtClient      apiextclientv1.CustomResourceDefinitionsGetter
 	MeteringClient    meteringclient.MeteringV1Interface
 	OLMV1Client       olmclientv1.OperatorsV1Interface
 	OLMV1Alpha1Client olmclientv1alpha1.OperatorsV1alpha1Interface
@@ -76,7 +76,7 @@ func New(logger logrus.FieldLogger, runLocal, runDevSetup bool, nsPrefix, repoDi
 		return nil, fmt.Errorf("failed to initialize the k8s clientset: %v", err)
 	}
 
-	apiextClient, err := apiextclientv1beta1.NewForConfig(config)
+	apiextClient, err := apiextclientv1.NewForConfig(config)
 	if err != nil {
 		return nil, fmt.Errorf("failed to initialize the apiextensions clientset: %v", err)
 	}